fix comments route

+++ src/api/blogger/api/blogs.controller.ts
@@ -73,10 +73,10 @@ export class BlogsController {
 	@HttpCode(204)
 	@Delete(':id')
 	async deleteBlogById(@Param('id') id: string, @Req() req: Request) {
-		const foundedBlog = await this.blogsService.findBlogById(id);
-		// if (foundedBlog.bloggerInfo.id !== req.user.id) {
-		// 	throw new ForbiddenException();
-		// }
+		const foundedBlog = await this.blogsService.findBlogByIdWithBloggerInfo(id);
+		if (foundedBlog.bloggerInfo.id !== req.user.id) {
+			throw new ForbiddenException();
+		}
 		const deletedBlog = await this.blogsService.deleteBlogById(id);
 		if (!deletedBlog) {
 			throw new NotFoundException();
@@ -92,13 +92,13 @@ export class BlogsController {
 		@Param('postId') postId: string,
 		@Req() req: Request,
 	) {
-		const foundedBlog = await this.blogsService.findBlogById(blogId);
+		const foundedBlog = await this.blogsService.findBlogByIdWithBloggerInfo(blogId);
 		if (!foundedBlog) {
 			throw new NotFoundException();
 		}
-		// if (foundedBlog.bloggerInfo.id !== req.user.id) {
-		// 	throw new ForbiddenException();
-		// }
+		if (foundedBlog.bloggerInfo.id !== req.user.id) {
+			throw new ForbiddenException();
+		}
 		const deletedPost = await this.blogsService.deletePostForSpecifiedBlog(postId, blogId);
 		if (!deletedPost) {
 			throw new NotFoundException();
@@ -110,13 +110,13 @@ export class BlogsController {
 	@HttpCode(204)
 	@Put(':id')
 	async updateBlogById(@Body() dto: UpdateBlogDto, @Param('id') id: string, @Req() req: Request) {
-		const foundedBlog = await this.blogsService.findBlogById(id);
+		const foundedBlog = await this.blogsService.findBlogByIdWithBloggerInfo(id);
 		if (!foundedBlog) {
 			throw new NotFoundException();
 		}
-		// if (foundedBlog.bloggerInfo.id !== req.user.id) {
-		// 	throw new ForbiddenException();
-		// }
+		if (foundedBlog.bloggerInfo.id !== req.user.id) {
+			throw new ForbiddenException();
+		}
 		await this.blogsService.updateBlogById(id, dto.name, dto.description, dto.websiteUrl);
 		return;
 	}
@@ -130,13 +130,13 @@ export class BlogsController {
 		@Param('postId') postId: string,
 		@Req() req: Request,
 	) {
-		const foundedBlog = await this.blogsService.findBlogById(blogId);
+		const foundedBlog = await this.blogsService.findBlogByIdWithBloggerInfo(blogId);
 		if (!foundedBlog) {
 			throw new NotFoundException();
 		}
-		// if (foundedBlog.bloggerInfo.id !== req.user.id) {
-		// 	throw new ForbiddenException();
-		// }
+		if (foundedBlog.bloggerInfo.id !== req.user.id) {
+			throw new ForbiddenException();
+		}
 		return await this.blogsService.updatePostForSpecifiedBlog(postId, dto);
 	}
 }
